requirements.txt
Django==2.0.4
Jinja2==2.10
redis==2.10.6
redispy==3.0.0
requests==2.18.4
requests-cache==0.4.13
tokenize-uk==0.2.0
translitua==1.2.4
psycopg2==2.7.3.2
tqdm==4.23.0
python-dateutil==2.7.2
elasticsearch_dsl<6,>=5s
Js2Py==0.59
python-Levenshtein==0.12.0